Ajout upload add/edit/remove image news

+++ app/Controller/NewsController.php
@@ -10,12 +10,15 @@ class NewsController extends AppController {
 		if ($this->request->is('post')) {
 			$d = $this->request->data;
 			$d['News']['news_id'] = null;
-			if ($this->News->save($d, true, array('news_id', 'news_title', 'news_content', 'news_img', 'news_is_activated'))) {
-				$this->Session->setFlash("La news à bien été ajouté !", 'notif');
-				$this->redirect(array('controller' => 'news', 'action' => 'index', 'admin' => true));
-			} else {
-				$this->Session->setFlash("Un problème est survnu, réessayer !", 'notif', array('type' => 'error'));
-				$this->redirect($this->referer());
+			$d['News']['news_img'] = $this->upload_file($d['News']['news_img'], 'news');
+			if ($d['News']['news_img'] != false) {
+				if ($this->News->save($d, true, array('news_id', 'news_title', 'news_content', 'news_img', 'news_is_activated'))) {
+					$this->Session->setFlash("La news à bien été ajouté !", 'notif');
+					$this->redirect(array('controller' => 'news', 'action' => 'index', 'admin' => true));
+				} else {
+					$this->Session->setFlash("Un problème est survnu, réessayer !", 'notif', array('type' => 'error'));
+					$this->redirect($this->referer());
+				}
 			}
 		}
 	}
@@ -25,12 +28,17 @@ class NewsController extends AppController {
 		if ($this->request->is('post') || $this->request->is('put')) {
 			$d = $this->request->data;
 			$this->News->id = $data['News']['news_id'];
-			if ($this->News->save($d, true, array('news_title', 'news_content', 'news_img', 'news_is_activated'))) {
-				$this->Session->setFlash("La news à bien été édité !", 'notif');
-				$this->redirect(array('controller' => 'news', 'action' => 'index', 'admin' => true));
-			} else {
-				$this->Session->setFlash("Un problème est survenu, Réessayer !", 'notif', array('type' => 'error'));
-				$this->redirect($this->referer());
+			$d['News']['news_img'] = $d['News']['news_img']['size'] == 0 ? $data['News']['news_img'] : $this->upload_file($d['News']['news_img'], 'news');
+			if ($d['News']['news_img'] != false) {
+				if ($data['News']['news_img'] != $d['News']['news_img'])
+					$this->delete_file($data['News']['news_img'], 'news');
+				if ($this->News->save($d, true, array('news_title', 'news_content', 'news_img', 'news_is_activated'))) {
+					$this->Session->setFlash("La news à bien été édité !", 'notif');
+					$this->redirect(array('controller' => 'news', 'action' => 'index', 'admin' => true));
+				} else {
+					$this->Session->setFlash("Un problème est survenu, Réessayer !", 'notif', array('type' => 'error'));
+					$this->redirect($this->referer());
+				}
 			}
 		}
 		$this->request->data = $data;
@@ -38,9 +46,12 @@ class NewsController extends AppController {
 
 	public function admin_remove($id) {
 		$this->autoRender = false;
-		$this->News->delete($id);
-		$this->Session->setFlash("La news à bien été supprimer !", 'notif');
-		$this->redirect($this->referer());
+		$news = $this->News->find('first', array('conditions' => array('news_id' => $id)));
+		if ($this->delete_file($news['News']['news_img'], 'news')) {
+			$this->News->delete($id);
+			$this->Session->setFlash("La news à bien été supprimer !", 'notif');
+			$this->redirect($this->referer());
+		}
 	}
 
 	public function admin_activated($id) {
